fix: Errors introduced through overlapping changes
